[Internal] Fix test_iam::test_scim_error_unmarshall integration test (#743)
## Changes
The error message returned by the SCIM API when using an invalid filter
changed, breaking our integration test. This PR updates the test to
check for the type of error returned, which should be robust to error
message changes. The new error message, `InvalidFilter request failed`,
is worse in my opinion, so I will raise this with the identity team.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` run locally
- [ ] `make fmt` applied
- [ ] relevant integration tests applied