[Internal] Refactor ApiClient into `_BaseClient` and `ApiClient` (#785)
## Changes
`ApiClient` is also coupled to the `Config` object, which means that it
can't be used in situations where there is no config. For example, when
fetching OIDC endpoints, the user may not have a complete `Config`
instance yet. However, failures when requesting from those endpoints
should still be retried according to the SDK's retry policy.

To address this, I've split the ApiClient into `_BaseClient` and
`ApiClient`. `_BaseClient` is the core implementation of the client
without any dependency on the `Config`. This is similar to what @rauchy
did in the Java SDK to cut the dependency between the `ApiClient` and
`DatabricksConfig`. The `_BaseClient` can then be used when fetching
OIDC endpoint information.

This will be used in
https://github.com/databricks/databricks-sdk-py/pull/784 to support
retrying OAuth OIDC endpoint fetches.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` run locally
- [ ] `make fmt` applied
- [ ] relevant integration tests applied