[Fix] Include package name for external types when deserializing responses (#786)
## Changes
#683 is caused by a small bug in the template used to generate the
Python SDK. When referring to a class defined in a separate API package,
only the module is imported, not the exact class, so the generated code
needs to use the qualified name of the structure.

Resolved #683.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` run locally
- [ ] `make fmt` applied
- [ ] relevant integration tests applied