[Internal] Better isolate ML serving auth unit tests (#803)
## Changes

This PR better isolates ML serving auth tests by explicitly ignoring the
config file if any. This solves an issue for users running the tests in
an environment (e.g. their local environment) that already has a
`.databrickscfg` file.

Note: the solution is a little hacky and we should think about a better
way to communicate that the config file should be ignored.

## Tests

Verified that the unit tests succeed in an environment that has a
`.databrickscfg`.

- [x] `make test` run locally
- [x] `make fmt` applied
- [x] relevant integration tests applied