[Feature] Read streams by 1MB chunks by default. (#817)
## What changes are proposed in this pull request?

This PR changes the `_BaseClient` to read streams by chunks of 1MB by
default. 1MB was chosen as a good compromise between speed and memory
usage (see PR #319).

Note that this is not a new feature per se as it was possible to
configure chunk size on the returned `_StreamResponse` before calling
its read method. However, the functionality was not easy to discover and
led several users to experience memory issues. The new default behavior
is more defensive.

## How is this tested?

Added a few test cases to verify that streams are chunked as expected.

---------

Signed-off-by: Renaud Hartert <renaud.hartert@databricks.com>