[Internal] Update Jobs list function to support paginated responses (#896)
## What changes are proposed in this pull request?
Introduces logic in the extension for jobs ListJobs call. The extended
logic accounts for the new response format of API 2.2. API 2.1 format
returns all tasks and job_cluster for each job in the jobs list. API 2.2
format truncates tasks and job_cluster to 100 elements. The extended
ListJobs logic calls GetJob for each job in the list to populate the
full list of tasks and job_clusters.

The code consumes generator that is returned from `super().list` and it
produces generator as well.

## How is this tested?
Unit tests and manual tests. Manual tests were done in two modes: using
API 2.2 and using API 2.1. So this code is code is compatible with both
API versions.

## Notes
Copy of this other PR
https://github.com/databricks/databricks-sdk-py/pull/889

The formatter is inconsistent and the original author could make it
work.

---------

Co-authored-by: Giorgi Kikolashvili <giorgi.kikolashvili@databricks.com>