[Internal] Implement async token refresh (#893)
## What changes are proposed in this pull request?

This PR is a step towards enabling asynchronous refreshes of data plane
tokens.
This PR updates the existing `Refreshable` abstract token class to
support async token refresh.

Note: async refreshes are disabled at the moment and will be enabled in
a follow-up PR.

## How is this tested?

Added unit tests.

## Changelog
The changelog entry will be added when the feature is enabled.

NO_CHANGELOG=true