[Internal] Switch code formatter to Black (#900)
## What changes are proposed in this pull request?

This PR switches the code formatter from `yapf` to `black`. 

The rationale is that we want a code formatter that (i) leaves as little
space as possible for discussions around formatting personal
preferences, and (ii) increases the likelihood of producing code that is
consistently formatted across the python open-source ecosystem.

At this point in time, we believe that `black` does a better job than
`yapf` and other alternatives to address these two problems.

## How is this tested?

No additional tests.

NO_CHANGELOG=true