[Internal] Make model serving unit tests a little more hermetic (#906)
## What changes are proposed in this pull request?

This PR is a no-op that makes the model serving unit tests a little more
hermetic. The solution is not perfect and we will have to consider
better ways to enable hermetic tests in the future (which might involve
a refactoring of the `Config` object in SDK-Mod).

## How is this tested?

N/A

NO_CHANGELOG=true