[Feature] Fix Model Serving User Credentials threading scenarios (#907)
## What changes are proposed in this pull request?

Previously in order to get invokers tokens, we always looked at the
current thread's thread data. however this is not always guaranteed as
langchain is internally multi threading calls. Therefore this PR only
looks at the main thread's data to get the credentials. The Main Thread
is guaranteed to have invokers credentials as we set them initially in
the scoring server.

## How is this tested?

Added unit tests

---------

Signed-off-by: aravind-segu <aravind.segu@databricks.com>
Co-authored-by: Renaud Hartert <renaud.hartert@databricks.com>