[Fix] Add support for all error details types and fix potential unmarshalling error (#912)
## What changes are proposed in this pull request?

This PR refactors how `DatabricksError` processes error details to fully
support the Databricks Error specification. In particular, it provides a
clear mechanism to access known error details types while still enabling
users to access unknown error details.

The core of the PR is the addition of a new `ErrorDetails` type which
operates as the union of all error details type. It implements the fact
that an error response should only contain at most one instance of each
type.

The code is purposely structured to hide implementation details by
separating the exported error details types from their unmarshalling
logic. This arguably leads to code that might be more complicated that
it ought to be. We could consider further simplifying it in the future
(e.g. by relying on the corresponding proto unmarshallers).

## How is this tested?

100% unit tests coverage of new code + slight refactor of the test
suite.