Move core directories and files to databricks directory (#937)
## What changes are proposed in this pull request?

This PR moves all the "core" directories and files into the "databricks"
directory, which is meant to become the core library.

This PR also comments out some tests(specifically tests in
tests/integration/test_auth.py) that upload the current copy of the
Python SDK to the workspace and use that copy to run some notebook code.
The reason for commenting out these tests is that Databricks Connect
depends on the Python SDK in the workspace. However, Due to the
directory changes, the dependency path in Databricks Connect no longer
matches this change. As SDK Mod represents a complete architectural
change of the SDK Beta, we cannot simply use it in the DB connect, and
we need to update it accordingly.

## How is this tested?
Existing Tests.
NO_CHANGELOG=true