Add Waiters for long-running operations (#950)
## What changes are proposed in this pull request?

This PR adds waiters to long-running operations, which provides
functionality to wait until a terminating state has been reached for the
long-running operation.

## How is this tested?
I re-enabled one integration test that uses long-running operations. I
will re-enable the rest of the tests in the next PR, as some of them
first need mixins that are removed because we don't have waiters.
NO_CHANGELOG=true