Fix VectorSearchAPI query_index  (#961)
## What changes are proposed in this pull request?
A regression was introduced in v0.51 of the Python SDK affecting the
`query_index()` method of the `VectorSearchAPI`, causing the query
result to be dropped during deserialization. This PR fixes this by
correcting the deserialization logic and the type of the API response.

## How is this tested?
This matches the implementation from the last working version v0.50.

---------

Signed-off-by: Miles Yucht <miles@databricks.com>