Fix highlighting issue due to multi-segment imports in WorkspaceClient/AccountClient (#979)
## What changes are proposed in this pull request?

This PR fixes a reported highlighting problem with the way API clients
are imported in `WorkspaceClient/AccountClient`. Specifically, it now
imports the API clients using their module as reference instead of a
multi-segment reference.

See highlighting issue:


![image](https://github.com/user-attachments/assets/98829599-8ec4-4e4f-8a36-3f526bf0f493)

This PR regenerates the SDK documentation that wasn't generated in the
previous release.

## How is this tested?

Unit and integration tests + verified locally that the objects are now
properly identified.