Fix FilesExt failed to retry on server side error issue (#1009)
## What changes are proposed in this pull request?

**WHAT**
- This PR fixes the issue that the old retry mechanism failed to
identify the retriable errors.
- It has done that by providing a new mechanism to support
distinguishing retriable and non-retriable errors solely by its status
code, which is the expected behavior when calling
_retry_cloud_idempotent_operation.
- The old _retry_idempotent_operation is removed because it is out of
use and can be misleading.
- The PR also refactors the `test_files`, which are the unit tests for
the `FilesExt` class.
  - Added typing for all methods and argument lists
- Introduced `UploadTestCase` as the base class for upload test cases,
in order to reduce the code duplication
- Renamed all `chunk` to `part`, to align the terminology with the HTTP
API.

**WHY**

- The previous implementation utilizes error mappers that's designed for
calling Databricks APIs (aka internal APIs), which the retry functions
relies on to determine whether the requests is retriable or not.
However, when calling the uploading API, it could be calling a CSP API
(aka external API), which we wish to distinguish whether it is retriable
solely on it's status code.
- Refactor of unit test file is done to improve its readability and
maintainability.

## How is this tested?

Unit tests.

**ALWAYS ANSWER THIS QUESTION:** Answer with "N/A" if tests are not
applicable
to your PR (e.g. if the PR only modifies comments). Do not be afraid of 
answering "Not tested" if the PR has not been tested. Being clear about
what
has been done and not done provides important context to the reviewers.