Fix slow unit tests for FilesExt (#1074)
## What changes are proposed in this pull request?
The unit tests for the FilesExt module are very slow because they
exercise the retry pathway using a real clock, resulting in long periods
spent waiting for retries. For unit tests, this is not necessary; they
can use the clock provided in the Config, which is a RealClock by
default in production but can be mocked in tests using FakeClock.

To fix this, I changed the call of `retried` to use the config's clock,
and then I updated the mock Config we provide as a fixture to use the
fake clock.

## How is this tested?
Existing tests should continue to work & pass.
Unit tests should be substantially faster than before.

NO_CHANGELOG=true