Revert interface change of FilesExt.upload (#1092)
## What changes are proposed in this pull request?

Provide the readers and reviewers with the information they need to
understand
this PR in a comprehensive manner. 

Specifically, try to answer the two following questions:

- **WHAT**
- Fix the issue where `FilesExt.upload`'s second parameter was changed
from `contents` to `content` unintentionally. This PR reverts the change
to make it compatible with the interface prior than version 0.68.0.
- **WHY**
- 0.69.0 introduced a breaking change unintentionally, which breaks the
usage for certain users. e.g. #1091


## How is this tested?

The unit tests has been updated to check the parameter name.