Fix FilesExt upload fails when content size is zero (#1088)
## What changes are proposed in this pull request?

Provide the readers and reviewers with the information they need to
understand
this PR in a comprehensive manner. 

Specifically, try to answer the two following questions:

- **WHAT** 
- Fix the `FilesExt` issue where it throws exception when the content
size is 0, by checking if the size of the content is below a certain
threshold. If it is below the threshold, upload using single-shot
instead of multipart upload.
- **WHY** 
  - `upload` interface needs to support upload data with zero size.

## How is this tested?

Unit test cases are updated to test the behavior change.